requirements.txt
gymnasium[classic-control]
opencv-python
rtamt
lark
pyyaml